🚑️ Handle `Hello` from peer upfront
Immediately after the connection is established, handle the `Hello` method
already. Otherwise, due to the async nature of our machinery, we may end
up sending other messages (e.g the `NameAcquired` signal) before the
return value of `Hello` is processed, which is unexpected.

This not only fixes the new mutli-connection test we just added but also
makes us pass most of the tests in zbus repo (except for 2).