🏗️ Complete re-work of name change signals
Mainly we don't use tokio MPSC channels for name change notifications
anymore. It turns out a successful send doesn't imply immediate reception
on the receiving task.

Moreover, the current implemenation was full of deadlocks, which these
changes fix.

If you're looking into this commit and thinking that it's not as atomic
as it should be, you're right. However, keeping in mind:

* the project is still very much in its infancy.
* there are 0 reviewers of my code.
* it took me weeks to finally get these changes working reliably.
* these changes finally enable all but 2 of zbus tests to pass.

I'm just going to make an exception here.