✅ tests: add config tests
The tests are from dbus reference implementation. The schema/dtd is a
based on a mix of what is accepted by dbus and dbus-broker.

Some tests are left with FIXME, as they may not fit in the config parser
itself, but rather in the runtime handling of options.

dbus has extra tests which may require installation to run (*.conf.in).