🏗️ Move the DBus interface back on our self-dial connection
We're going to implemementing the Monitoring interface soon and I don't
see anothher not-too-complicated way to get our own ObjectServer replies
delivered to monitor peers.

Besides the hangs I was seeing (the main reason I gave up on this
before) turned out to be unrelated and were caused by us trying to take
a lock multiple times in the same task/thread.