✨ Cancel peer's stream when it's dropped
Currently this isn't needed because when a peer is removed, its message
stream and associated tasks gets automatically dropped because the
connection also goes away.

However, when we implement Monitoring interface, we'll need to keep the
connection around while the dropping the peer and its associated message
stream and task.