🛂 Drop auth-mechanism CLI arg
It's now automatically chosen for you, based on the socket type in use.
For UNIX sockets and for TCP on Windows, we use `EXTERNAL`. For TCP
sockets on non-Windows, we use `ANONYMOUS` (i-e no authentication at all).

On Unix machines, people should use the UNIX socket anyway and if
EXTERNAL is possible, it's best to stick to it. `dbus-broker` also
doesn't support anonymous authentication and since they only support
Unix sockets for transport, it means they don't support anonymous
authentication at all.