♻️ Extract `unix_addr()` out of `unix_stream()`
Previously, we'd take the `--address` or default address, resolve it
into the socket address, and start listening on that address. But we
need to be able to expose that address to clients, too. So we break
`Bus::unix_stream()` apart giving us a separate `Bus::unix_addr()` and
access to the intermediate socket address.