⚡️ More binary optimizations for release builds (#228)
These optimizations reduce the binary size by 40%, which is 3.4M now.

This change also means that we now strip the binary of debug symbols and
generate a `busd.dwp` (in the target dir) so that debug symbols are
still available when/where needed. Distro packagers should take a note
of this.